+++ examples/ESP_WiFi_MQTT/dynamicParams.h
@@ -31,6 +31,15 @@
   char *pdata;
   uint8_t maxlen;
   } MenuItem;
+
+  The ids must be unique. An id starting with * will be obscured by type password.
+  Ids starting with _ will not use an HTML value tag and need to be reentered every time.
+  Please be noted that the following reserved names are already used in library:
+    "id"    for WiFi SSID
+    "pw"    for WiFi PW
+    "id1"   for WiFi1 SSID
+    "pw1"   for WiFi1 PW
+    "nm"    for Board Name
 **************************************/
 
 #if USE_DYNAMIC_PARAMETERS
@@ -57,7 +66,7 @@ MenuItem myMenuItems [] =
   { "svr", "AIO_SERVER",      AIO_SERVER,     AIO_SERVER_LEN },
   { "prt", "AIO_SERVERPORT",  AIO_SERVERPORT, AIO_SERVERPORT_LEN },
   { "usr", "AIO_USERNAME",    AIO_USERNAME,   AIO_USERNAME_LEN },
-  { "key", "AIO_KEY",         AIO_KEY,        AIO_KEY_LEN },
+  { "*key", "AIO_KEY",        AIO_KEY,        AIO_KEY_LEN },
   { "pub", "AIO_PUB_TOPIC",   AIO_PUB_TOPIC,  AIO_PUB_TOPIC_LEN },
   { "sub", "AIO_SUB_TOPIC",   AIO_SUB_TOPIC,  AIO_SUB_TOPIC_LEN },
 };